Move StateDB into its own file to facilitate testing.
